+++ src/main/java/com/saasquatch/sdk/auth/BasicAuth.java
@@ -1,7 +1,8 @@
 package com.saasquatch.sdk.auth;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
-import org.apache.commons.codec.binary.Base64;
+
+import net.iharder.Base64;
 import org.apache.hc.client5.http.async.methods.SimpleRequestBuilder;
 import org.apache.hc.core5.http.HttpHeaders;
 
@@ -18,7 +19,7 @@ final class BasicAuth implements AuthMethod {
   @Override
   public void mutateRequest(SimpleRequestBuilder requestBuilder) {
     requestBuilder.setHeader(HttpHeaders.AUTHORIZATION,
-        "Basic " + Base64.encodeBase64String((username + ':' + password).getBytes(UTF_8)));
+        "Basic " + Base64.encodeBytes((username + ':' + password).getBytes(UTF_8)));
   }
 
 }