+++ src/main/java/com/saasquatch/sdk/internal/InternalUtils.java
@@ -3,9 +3,7 @@ package com.saasquatch.sdk.internal;
 import static com.saasquatch.sdk.internal.json.GsonUtils.gson;
 import static java.nio.charset.StandardCharsets.UTF_8;
 
-import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
+import com.google.gson.JsonSyntaxException;
 import com.saasquatch.sdk.exceptions.SaaSquatchApiException;
 import com.saasquatch.sdk.exceptions.SaaSquatchUnhandledApiException;
 import com.saasquatch.sdk.input.UserIdInput;
@@ -20,6 +18,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.UnsupportedEncodingException;
 import java.net.URLDecoder;
+import java.net.URLEncoder;
 import java.nio.CharBuffer;
 import java.nio.charset.Charset;
 import java.util.AbstractMap.SimpleImmutableEntry;
@@ -42,8 +41,7 @@ import java.util.zip.GZIPInputStream;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import javax.annotation.WillNotClose;
-import org.apache.commons.codec.binary.Base64;
-import org.apache.commons.codec.net.URLCodec;
+import net.iharder.Base64;
 import org.apache.hc.client5.http.async.methods.SimpleHttpRequest;
 import org.apache.hc.client5.http.async.methods.SimpleHttpResponse;
 import org.apache.hc.client5.http.impl.async.CloseableHttpAsyncClient;
@@ -244,7 +242,12 @@ public final class InternalUtils {
    * RFC3986 URL encode
    */
   public static String urlEncode(@Nonnull String s) {
-    return new String(URLCodec.encodeUrl(RFC_3986_SAFE_CHARS, s.getBytes(UTF_8)), UTF_8);
+    try {
+      return URLEncoder.encode(s, UTF_8.name())
+          .replace("+", "%20").replace("*", "%2A").replace("%7E", "~");
+    } catch (UnsupportedEncodingException e) {
+      throw new RuntimeException(e); // Won't happen
+    }
   }
 
   /**
@@ -362,11 +365,21 @@ public final class InternalUtils {
       if (result == null) {
         return null;
       }
-      @SuppressWarnings("unchecked") final Map<String, Object> resultAsMap =
-          (Map<String, Object>) result;
-      result = resultAsMap.get(key);
+      //noinspection unchecked
+      result = ((Map<String, Object>) result).get(key);
     }
     return result;
+  }
+
+  public static String addBase64Padding(String base64) {
+    if (base64.length() % 4 == 0) {
+      return base64; // Avoid unnecessary copying with StringBuilder
+    }
+    final StringBuilder base64Builder = new StringBuilder(base64);
+    do {
+      base64Builder.append('=');
+    } while (base64Builder.length() % 4 != 0);
+    return base64Builder.toString();
   }
 
   /**
@@ -378,15 +391,19 @@ public final class InternalUtils {
       throw new IllegalArgumentException("Invalid JWT");
     }
     final String payloadPart = jwtParts[1];
-    // Do not use the overload that takes a String. It does not work on Android.
-    final byte[] payloadBytes = Base64.decodeBase64(payloadPart.getBytes(UTF_8));
-    final JsonElement jsonElement = JsonParser.parseString(new String(payloadBytes, UTF_8));
-    if (!(jsonElement instanceof JsonObject)) {
-      throw new IllegalArgumentException("JWT payload is not a JSON object");
+    final byte[] payloadBytes;
+    // This Base64 library expects the base64 string to have proper padding
+    try {
+      payloadBytes = Base64.decode(addBase64Padding(payloadPart), Base64.URL_SAFE);
+    } catch (IOException e) {
+      throw new IllegalArgumentException("Invalid JWT payload", e);
     }
-    @SuppressWarnings("unchecked") final Map<String, Object> payloadMap =
-        gson.fromJson(jsonElement, Map.class);
-    return payloadMap;
+    try {
+      //noinspection unchecked
+      return gson.fromJson(new String(payloadBytes, UTF_8), Map.class);
+    } catch (JsonSyntaxException e) {
+      throw new IllegalArgumentException("Invalid JWT payload", e);
+    }
   }
 
   @Nonnull